+++ packages/autocomplete/src/component/Autocomplete.jsx
@@ -50,7 +50,9 @@ type Props = {
   newEntryText?: string;
   onNewEntry?: (label: string, callback: Function) => Promise<Item>;
   /** Flyout direction */
-  direction?: "left" | "right"
+  direction?: "left" | "right";
+  required?: boolean;
+  placeholder?: string;
 };
 
 type IState = {
@@ -68,7 +70,7 @@ class Autocomplete extends Component<Props, IState> {
     cursor: 0,
     selection: [],
     defaultValue: this.props.defaultValue || ''
-  }
+  };
 
   inputField: HTMLInputElement;
   selectionMode = this.props.multipleSelect ? new MultipleSelectionMode(this) : new SingleSelectionMode(this);
@@ -96,7 +98,7 @@ class Autocomplete extends Component<Props, IState> {
       takeUntil(this.destroy$),
       tap(e => {
         e.preventDefault();
-        const {results, cursor} = this.state
+        const {results, cursor} = this.state;
         if (e.key === "ArrowDown" && cursor < results.length - 1) {
           this.setState({
             open: true,
@@ -124,7 +126,8 @@ class Autocomplete extends Component<Props, IState> {
     );
 
     // Start the show!
-    this.selectionMode.handleDefaultValue(this.props.defaultValue)
+    this.searchMode.initialize();
+
     handleArrowKeys$.subscribe();
     change$.subscribe();
   }
@@ -165,7 +168,7 @@ class Autocomplete extends Component<Props, IState> {
   scrollToItem = () => {
     const domNode = ReactDOM.findDOMNode(this['item_' + this.state.cursor]);
     domNode.scrollIntoView(false)
-  }
+  };
 
   selectOption(item: Item) {
     if (!!item)
@@ -185,7 +188,7 @@ class Autocomplete extends Component<Props, IState> {
         {item.label}
       </li>
     );
-  }
+  };
 
   componentWillUnmount() {
     this.destroy$.next();
@@ -199,11 +202,12 @@ class Autocomplete extends Component<Props, IState> {
              autoComplete="off"
              disabled={this.props.disabled}
              ref={ref => this.inputField = ref}
-             defaultValue={this.state.defaultValue}
              data-qa={this.props.qa}
              {...this.props.state}
              onBlur={() => this.closePane()}
              onFocus={() => this.openPane()}
+             placeholder={this.props.placeholder || ''}
+             required={!!this.props.required}
       />
     );
   }
@@ -237,6 +241,7 @@ class Autocomplete extends Component<Props, IState> {
         'has-icon-right': this.props.loading || this.state.isLoading,
         'has-icon-left': this.props.showSearchIcon,
         [`${stateClasses[state]}`]: !!state,
+        'is-required': !!this.props.required
       }
     );
 
@@ -245,7 +250,7 @@ class Autocomplete extends Component<Props, IState> {
       {
         'has-focus': !!this.inputField && this.inputField === document.activeElement && !!this.props.multipleSelect
       }
-    )
+    );
 
     const newEntryClasses = classNames({
       'm-selectable-list__item': true,
@@ -259,7 +264,7 @@ class Autocomplete extends Component<Props, IState> {
             <InputLabel htmlFor={this.props.id}>{this.props.label}</InputLabel>
             <div className={wrapperClasses} onClick={() => this.focusOnInput()}>
               {this.props.multipleSelect && <TagList>
-                {this.state.selection.map(s => {
+                {this.state.selection.filter(s => !!s).map(s => {
                   return (
                     <TagListItem closable={true} onClick={() => this.selectionMode.unselect(s)} key={s.value}
                                  value={s.label}/>)